it works 👼
After many hours of debug, several mistakes, many beers deep we finally have something that talks with each other. For now only station and server talks with each other. By the great heroic effort of @HaugPixel and @Stian108 we got it working. It was @HaugPixel ‘s brilliant observation we discovered that the docker images did not rebuild with a simple docker compose up, as I had assumed. So effectively we had wasted 2 hours of debug because docker didn’t use the new python scripts we tried to debug. So we alter some environment settings and now docker flushes the python buffer every new line, so we know the server container works and if it is receiving data from the weather station.  Then we fixed Docker to not use TCP ports but forward UDP ports, the station now talks to a server hostname not the localhost, the server listens on all the addresses in its container not localhost hostname. All of this we figured out with the help of almighty @Stian108 powerful google-fu abilities and @HaugPixel hawkeyed bug vision. Now we need to do the other server…

Co-Authored-By: Erlend Haugen <69681461+HaugPixel@users.noreply.github.com>
Co-Authored-By: Circle Of Stian <15105930+stian108@users.noreply.github.com>